# specs/api-contract.yaml
openapi: 3.0.3
info:
  title: SmartView API v2
  description: ISP broadband monitoring dashboard — backend API contract
  version: 2.0.0
servers:
  - url: http://localhost:3001
    description: Local development
  - url: /api
    description: Docker (nginx proxy)
...
security:
  - cookieAuth: []

# Note: cookieAuth is defined in components/securitySchemes below.
# All data routes require the auth_token HttpOnly cookie issued by POST /api/auth/login.
paths:
...
  /api/customers/{customerId}/devices:
    get:
      tags: [devices]
      summary: List all in-home devices for a customer with current stats
      security:
        - cookieAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
          example: "cust-uuid-1234"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated list of devices with current stats
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DeviceWithStats"
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 8
                      limit:
                        type: integer
                        example: 50
                      offset:
                        type: integer
                        example: 0
                      hasMore:
                        type: boolean
                        example: false
        "401":
          description: Not authenticated
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/customers/{customerId}/devices/{deviceId}/history:
    get:
      tags: [devices]
      summary: >
        Sampled historical stats for a device.
        Uses the same step-down sampling strategy as /modems/:customerId/history.
      security:
        - cookieAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          description: Number of days of history to return
          schema:
            type: integer
            enum: [1, 7, 30, 90]
            default: 7
        - name: limit
          in: query
          description: Maximum data points to return after sampling
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 200
      responses:
        "200":
          description: Sampled device history
          content:
            application/json:
              schema:
                type: object
                required: [data, samplingApplied]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DeviceHistoryRecord"
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Raw row count before sampling
                        example: 360
                      limit:
                        type: integer
                        example: 200
                      offset:
                        type: integer
                        example: 0
                      hasMore:
                        type: boolean
                        example: false
                  samplingApplied:
                    type: boolean
                    description: True when raw row count exceeded limit and step-down sampling was applied
                    example: true
        "401":
          description: Not authenticated
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── MODEM ──────────────────────────────────────────────────────────────────
...
components:
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Not found"
...
    DeviceType:
      type: string
      enum:
        - laptop
        - desktop
        - phone
        - tablet
        - games_console
        - smart_tv
        - wifi_extender
        - router
        - iot
        - other
    ConnectionType:
      type: string
      enum:
        - ethernet
        - wifi_2_4
        - wifi_5
        - wifi_6
    Device:
      type: object
      required: [id, customer_id, name, device_type, connection_type, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        customer_id:
          type: string
          example: "cust-uuid-1234"
        parent_device_id:
          type: string
          nullable: true
          description: ID of the upstream device this one connects through. NULL for root gateway.
          example: "router-uuid-5678"
        name:
          type: string
          example: "Living Room TV"
        device_type:
          $ref: "#/components/schemas/DeviceType"
        connection_type:
          $ref: "#/components/schemas/ConnectionType"
        mac_address:
          type: string
          nullable: true
          example: "aa:bb:cc:dd:ee:ff"
        created_at:
          type: string
          format: date-time
          example: "2026-01-15T10:30:00Z"
...
    DeviceWithStats:
      description: Device metadata joined with current stats snapshot
      allOf:
        - $ref: "#/components/schemas/Device"
        - type: object
          properties:
            is_online:
              type: integer
              enum: [0, 1]
              example: 1
            rssi_dbm:
              type: number
              nullable: true
              example: -62.5
            upload_mbps:
              type: number
              example: 12.4
            download_mbps:
              type: number
              example: 87.3
            latency_ms:
              type: number
              example: 4.2
            stats_recorded_at:
              type: string
              format: date-time
              example: "2026-04-19T08:00:00Z"
    DeviceHistoryRecord:
      type: object
      required: [id, device_id, is_online, upload_mbps, download_mbps, latency_ms, recorded_at]
      properties:
        id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        is_online:
          type: integer
          enum: [0, 1]
        rssi_dbm:
          type: number
          nullable: true
        upload_mbps:
          type: number
        download_mbps:
          type: number
        latency_ms:
          type: number
        recorded_at:
          type: string
          format: date-time
...